pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "vikunja-mcp-server"
version = "0.1.0"
description = "Production-oriented MCP server and orchestration helpers for Vikunja"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Vikunja MCP Contributors" }]
dependencies = [
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.1",
  "tenacity>=8.4.0",
  "typer>=0.12.3",
  "mcp>=1.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
]

[project.scripts]
vikunja-mcp = "vikunja_mcp.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/vikunja_mcp"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["B008"]
